testing CrossEntropyLoss

+++ train_EEG.py
@@ -609,7 +609,7 @@ def run_train_EC_Multi(learn_rate, wd, batch_sz, epochs, outfile):
 
             print(f"lable: {label}")
             label_reshaped = np.reshape(label, (batch_sz, 5))
-            label_reshaped = label_reshaped.type(torch.float32)
+            label_reshaped = label_reshaped.type(torch.long)
             print(f"reshaped: {label_reshaped}")
 
             test = []
@@ -630,7 +630,7 @@ def run_train_EC_Multi(learn_rate, wd, batch_sz, epochs, outfile):
                 output, h_res = my_mental.forward(p, n_entry, h_1)
                 h = h_res
 
-            output = output.type(torch.float32)
+            #output = output.type(torch.float32)
         
             loss = torch.nn.CrossEntropyLoss()
             res = loss(output, label_reshaped)
